+++ Bitmex.Net/Bitmex.Net.Client.csproj
@@ -31,7 +31,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="CryptoExchange.Net" Version="5.1.11" />
+    <PackageReference Include="CryptoExchange.Net" Version="5.3.0" />
     <PackageReference Include="CsvHelper" Version="27.1.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
